feat: add hook for home page to get and filter products
