+++ src/pages/home/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useCategoryCarousel } from "./use-category-carousel.ts";
+export { useProductFilters } from "./use-products-filters.ts";
